+++ tests/public-api/test_solvers_jobs_api.py
@@ -14,7 +14,6 @@ import logging
 import time
 from operator import attrgetter
 from pathlib import Path
-from typing import Any
 from urllib.parse import quote_plus
 from zipfile import ZipFile
 
@@ -23,6 +22,8 @@ import osparc.exceptions
 import pytest
 from osparc import FilesApi, SolversApi
 from osparc.models import File, Job, JobInputs, JobOutputs, JobStatus, Solver
+from pytest import TempPathFactory
+from pytest_simcore.helpers.utils_public_api import ServiceInfoDict, ServiceNameStr
 from tenacity import Retrying, TryAgain
 from tenacity.after import after_log
 from tenacity.retry import retry_if_exception_type
@@ -39,7 +40,7 @@ logger = logging.getLogger(__name__)
 @pytest.fixture(scope="module")
 def sleeper_solver(
     solvers_api: SolversApi,
-    services_registry: dict[str, Any],
+    services_registry: dict[ServiceNameStr, ServiceInfoDict],
 ) -> Solver:
     # this part is tested in test_solvers_api so it becomes a fixture here
 
@@ -84,12 +85,12 @@ def sleeper_solver(
 
 
 @pytest.fixture(scope="module")
-def uploaded_input_file(tmpdir_factory, files_api: FilesApi) -> File:
+def uploaded_input_file(tmp_path_factory: TempPathFactory, files_api: FilesApi) -> File:
 
-    tmpdir = tmpdir_factory.mktemp("uploaded_input_file")
+    basedir: Path = tmp_path_factory.mktemp("uploaded_input_file")
 
     # produce an input file in place
-    input_path = Path(tmpdir) / "file-with-number.txt"
+    input_path = basedir / "file-with-number.txt"
     input_path.write_text("2")
 
     # upload resource to server
